requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pkg-resources==0.0.0
PyYAML==3.12
requests==2.12.1
Werkzeug==0.11.11
